requirements.txt
streamlit

langchain
langchain-community
openai
pinecone-client
PyPDF2